+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 ## 0.0.4
 
 - fix not correctly closing all files
+
+## 0.1.0
+
+- better api
+- add new command to close opened file siblings
+- remove close non dirty cmnd as its natively supported by the editor